+++ lib/charms/getTransaction.ts
@@ -1,5 +1,5 @@
 import { prisma } from '@charmverse/core/prisma-client';
 
-export async function getCharmTx(id: string) {
+export async function getTransaction(id: string) {
   return prisma.charmTransaction.findFirst({ where: { id } });
 }
