+++ pages/api/profile/charms/history.ts
@@ -1,12 +1,8 @@
-import { InvalidInputError } from '@charmverse/core/errors';
 import type { CharmTransaction } from '@charmverse/core/prisma-client';
 import type { NextApiRequest, NextApiResponse } from 'next';
 import nc from 'next-connect';
 
-import type { CharmTxResult } from 'lib/charms/addCharmTransaction';
-import { getUserOrSpaceWallet } from 'lib/charms/getUserOrSpaceWallet';
 import { listTransactionsHistory } from 'lib/charms/listTransactionsHistory';
-import { transferCharms } from 'lib/charms/transferCharms';
 import { onError, onNoMatch, requireUser } from 'lib/middleware';
 import { withSessionRoute } from 'lib/session/withSession';
